chore: update wallet concept story
