+++ components/brave_wallet_ui/common/async/__mocks__/lib.ts
@@ -5,6 +5,7 @@
 
 import { mockBasicAttentionToken, mockEthToken } from '../../../stories/mock-data/mock-asset-options'
 import { BraveWallet, GetChecksumEthAddressReturnInfo, GetEthAddrReturnInfo } from '../../../constants/types'
+import { HardwareWalletConnectOpts } from '../../../components/desktop/popup-modals/add-account-modal/hardware-wallet-connect/types'
 
 let mockedAllowance = '1000000000000000000' // 1 unit
 let mockedIsSwapSupported = true
@@ -87,4 +88,25 @@ export const isStrongPassword = (value: string) => {
     value.toLowerCase() !== value && // contains an uppercase character
     /\d/.test(value) // contains a number
   )
+}
+
+export const onConnectHardwareWallet = (opts: HardwareWalletConnectOpts): Promise<BraveWallet.HardwareWalletAccount[]> => {
+  const makeDerivationPath = (index: number): string => `m/44'/60'/${index}'/0/0`
+
+  return new Promise((resolve) => {
+    resolve(Array.from({ length: opts.stopIndex - opts.startIndex }, (_, i) => ({
+      address: '0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2',
+      derivationPath: makeDerivationPath(i + opts.startIndex),
+      name: 'Ledger 1',
+      hardwareVendor: 'Ledger',
+      deviceId: 'device1',
+      coin: BraveWallet.CoinType.ETH
+    })))
+  })
+}
+
+export const getBalance = (): Promise<string> => {
+  return new Promise(async (resolve) => {
+    resolve('0')
+  })
 }
